Fixed issue with hide date but no due date
Behaviour will be as below:
If due is not available and hide is provided with a negative
short format modifier like -2, -5 etc, then don't assign a hide
date.
But if the short modifier is a positive like +2 or +5 then
calculate relative to current date and store the hide date.

Additional changes:
- Single Event_ID for all tasks modified via a single 'modify'
operation to enable the future functionality of 'undo'.
- Commented a few print lines