Loads of small changes
1. Removed event_stack table and bin tables. Will now be managed
within the workspace tables

2. Task operations are now along with add and modify:
	START - Task status changes to STARTED
	STOP - Task status changes to TO_DO
	DONE - Task status changes to DONE and moves into completed
	REVERT - Task status changes to TO_DO and moves into pending
This is to better align with time tracking functions that are planned

3. Updated get_tags function to return a comma seprated string of
tags or None. This is instead of the old logic of returning a list
and duplication of conversion to string repeated in multiple
functions

4. On any non ADD operation after the updated version is inserted
all prior version have their task ID set to '-'. This ensures at any
point only the latest version of tasks in the 'pending' area will have a
task ID