Added some logging and fixed quite a few issues
1. Added logging using python's native logging and a new option for
all commands to allow DEBUG messages to be shown -v or --verbose
Default logging level is ERROR.

2. Added a check where only the high level area filters are provided
and in such cases check for user confirmation. This is done by
assigning a new key value pair in parse_filters as 'HL_FILTERS_ONLY'
and using this in the operations functions to get user confirmation.

3. Set the colouring for bin tasks same as done tasks in display_tasks

4. Reordered the if check in display_tasks so that the BIN and DONE
formatting gets applied

5. Update get_and_printactive_task_count to get_and_print_task_count
and takes in an additional parameter for area. This function now
prints counts in DONE and BIN areas and is controlled by the new
area parameter

6. Added additional tests to include the start, stop, revert, delete
and empty functions.

7. Uses mock library now to allow for prompt input to the pytests.
Updated setup.py accordingly.