Introduced priority, custom themes and comments and code reorg
1. Introduced a priority attribute for the tasks with High, Medium,
Normal and Low. This can be set as part of task add and modify.
Default is Normal. Added relevant code in add_task, modify_task
and added a new function to translate the priority from a user
provided value. Updated ORM Workspace class as well.

2. Added custom themes for Rich Library to make maintenance
easier.

3. Replaced click's echo with Console's Print except for scenario
noted below. For exception replaced it with LOGGER.debug. As
Click's testing library is used where the test depend on the
output a context manager is used to redirect the CONSOLE.print()
output to click's echo. This is used for example in
reflect_object_n_print()
The CONSOLE.print() is now used to format messages which indicate
change done by the user's actions.

4. Added an option to use pager output for the view command
using an additional user provided filter of '^'
The pager is invoked using a context manager for CONSOLE.print.

5. Where the entire ORM object needs to be printed, using
the inspect function from sqlalchmey. This avoids having to
hardcode the column names in the print function. Added a
global list which indicates which fields need to be printed
in a non error/debug scenario, ie for user output.

6. Added a hybrid property in Workspace ORM class to provide
the difference between the due date and today. This is used
as of now in display_tasks() where addl_info virtual column has been
changed to'due in' to give the number of days the task will be
due in. When trying to cast the date difference as an integer
from a float and then conact with the string 'DAY(S)' the
library translates this to a add operation rather than concat.
Instead used substr and instr to remove the decimal part of
the datediff.
Additional note on 'addl_info'/due in', now task in completed
and bin areas will show this column as '-'

7. Removed the redundant columns for 'done' and 'modified'.
The column 'created' can be used to identify both these
due to the nature of versioning for every operation.

8. Modified tests to match modified output for the commands
related to point 3 above.