Additional changes related to recurring tasks
1. Added required logic for the done operation to deal with recurring
tasks. 2 additional pieces are added, one to create a derived task when
there are no more derived tasks in the pending area. This is to give a
task entry for the users to modify if required. The other is to move the
base task to the completed area when the last derived task is moved into
the completed area