Split of revert into revert/reset and bug fix on using shortform for end
- The revert command which worked on either tasks in pending to reset
duration to 0 or for task in completed to reset duration to 0 and move
them to pending has been changed. Now revert will be used solely to move
tasks from completed to pending, ie from DONE to TO_DO and retaining any
duration it had earlier. To reset duration a new command has been
included, 'reset'. This will work on only pending tasks and will change
status of a task to TO_DO and update the duration to 0.

- Encountered a bug where when using modify with a shorform date for end
the app does not do conversion. This has been fixed.