Fixed duration carry forward in now and duration calc when > 1day
- When now command is run the duration was not being carried forward for
the task which was having its now flag reset. This was due to NOW now
being included in the condition to calculate the duration in
cal_duration()

- When the duration to be calculated on STOP operation was more than a
day the duration was incorrect. This was due to the use of the 'seconds'
attribute from  a time difference expression rather than
'total_seconds'. This change is again in calc_duration()